requirements.txt
streamlit
openai
requests
python-dotenv
opencv-contrib-python 
numpy 
Pillow
python-Levenshtein